+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.2.3 (Status: DEVELOPMENT)
 * nafsdm-master now uses SQL to save domains (updated nafsdmctl and slave with the new save format)
+* nafsdmctl list now prints in a nicer format
 
 Version 1.2.2 (Status: RELEASED)
 * Reworked logger function completely (now catches exceptions, new logging levels, stdout and filehandler)